Feature/flow usecase

+++ data/src/main/java/com/ptut/data/entity/MovieData.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 data class MovieData(
     val adult: Boolean,
     val backdropPath: String,
-    val genreIds: List<Int>,
+    val genreIds: List<Long>,
     val id: Long,
     val originalLanguage: String,
     val originalTitle: String,
